+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Remove `#features` and `#routes` writer and `#configuration` reader from `Rodauth::Rails::Auth` (@janko)
+
 ## 0.13.0 (2021-06-10)
 
 * Add `:query`, `:form`, `:session`, `:account`, and `:env` options to `Rodauth::Rails.rodauth` (@janko)